+++ docs/chain.contracts.rst
@@ -1,0 +1,6 @@
+.. _chain-contracts:
+
+
+Introduction to Chains
+======================
+
