+++ docs/chain.provider.rst
@@ -1,0 +1,1 @@
+.. _chain-provider:
