+++ docs/chain.store.rst
@@ -1,0 +1,1 @@
+.. _chain-store:
