+++ docs/populus.chain.rst
@@ -1,0 +1,54 @@
+populus.chain package
+=====================
+
+Submodules
+----------
+
+populus.chain.base module
+-------------------------
+
+.. automodule:: populus.chain.base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.chain.external module
+-----------------------------
+
+.. automodule:: populus.chain.external
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.chain.geth module
+-------------------------
+
+.. automodule:: populus.chain.geth
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.chain.tester module
+---------------------------
+
+.. automodule:: populus.chain.tester
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.chain.testrpc module
+----------------------------
+
+.. automodule:: populus.chain.testrpc
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: populus.chain
+    :members:
+    :undoc-members:
+    :show-inheritance:
