+++ docs/populus.config.upgrade.rst
@@ -1,0 +1,22 @@
+populus.config.upgrade package
+==============================
+
+Submodules
+----------
+
+populus.config.upgrade.v1 module
+--------------------------------
+
+.. automodule:: populus.config.upgrade.v1
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: populus.config.upgrade
+    :members:
+    :undoc-members:
+    :show-inheritance:
