+++ docs/populus.contracts.rst
@@ -1,0 +1,61 @@
+populus.contracts package
+=========================
+
+Subpackages
+-----------
+
+.. toctree::
+
+    populus.contracts.backends
+
+Submodules
+----------
+
+populus.contracts.contract module
+---------------------------------
+
+.. automodule:: populus.contracts.contract
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.contracts.exceptions module
+-----------------------------------
+
+.. automodule:: populus.contracts.exceptions
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.contracts.provider module
+---------------------------------
+
+.. automodule:: populus.contracts.provider
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.contracts.registrar module
+----------------------------------
+
+.. automodule:: populus.contracts.registrar
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+populus.contracts.store module
+------------------------------
+
+.. automodule:: populus.contracts.store
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: populus.contracts
+    :members:
+    :undoc-members:
+    :show-inheritance:
